output: Turn winsys_id into a uint64_t
We need a 64 bit field to combine the device id and connector
id into the output id (winsys_id).

+++ src/core/window-private.h
@@ -231,7 +231,7 @@ struct _MetaWindow
 
   double tile_hfraction;
 
-  int preferred_output_winsys_id;
+  uint64_t preferred_output_winsys_id;
 
   /* Whether we're shaded */
   guint shaded : 1;
